ar: rearrange network hook to support building on windows
