--- client/allocrunner/alloc_runner_nonlinux.go
@@ -1,13 +1,0 @@
-// +build !linux
-
-package allocrunner
-
-import (
-	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/nomad/client/allocrunner/interfaces"
-)
-
-// noop for non linux clients
-func (ar *allocRunner) initPlatformRunnerHooks(hookLogger hclog.Logger) ([]interfaces.RunnerHook, error) {
-	return nil, nil
-}
